Fixed Update and Delete

+++ budget-tracker-react-client/src/App.js
@@ -3,7 +3,6 @@ import Constants from "./utilities/Constants"
 import TransactionCreateForm from "./components/TransactionCreateForm"
 import TransactionUpdateForm from "./components/TransactionUpdateForm"
 import TransactionTable from "./components/TransactionTable";
-
 export default function App() {
   const [transactions, setTransactions] = useState([]);
   const [showingCreateNewTransactionForm, setShowingCreateNewTransactionForm] = useState(false);
@@ -18,7 +17,7 @@ export default function App() {
     })
       .then(response => response.json())
       .then(transactionsFromServer => {
-        setTransactions(transactionsFromServer)
+        setTransactions(transactionsFromServer);
       })
       .catch((error) => {
         console.log(error);
@@ -27,20 +26,22 @@ export default function App() {
   }
 
   function deleteTransaction(transactionId) {
-    const url = `${Constants.API_URL_DELETE_TRANSACTION_BY_ID}/${transactionId}`;
+    if (window.confirm(`Are you sure you want to delete?`)) {
+      const url = `${Constants.API_URL_DELETE_TRANSACTION_BY_ID}/${transactionId}`;
 
-    fetch(url, {
-      method: 'DELETE',
-    })
-      .then(response => response.json())
-      .then(responseFromServer => {
-        console.log(responseFromServer);
-        onTransactionDeleted(transactionId);
+      fetch(url, {
+        method: 'DELETE',
       })
-      .catch((error) => {
-        console.log(error);
-        alert(error);
-      });
+        .then(response => response.json())
+        .then(responseFromServer => {
+          console.log(responseFromServer);
+          onTransactionDeleted(transactionId);
+        })
+        .catch((error) => {
+          console.log(error);
+          alert(error);
+        });
+    }
   }
 
   return (
@@ -67,43 +68,16 @@ export default function App() {
     </div>
   );
 
-  function renderTransactionsTableOld() {
-    getTransactions();
-    return (
-      <div className="table-responsive mt-4">
-        <table className="table">
-          <thead>
-            <tr>
-              <th scope="col">Value</th>
-              <th scope="col">Note</th>
-              <th scope="col"></th>
-            </tr>
-          </thead>
-          <tbody>
-            {transactions.map((transaction) => (
-              <tr key={transaction.transactionId}>
-                <td>{transaction.value}</td>
-                <td>{transaction.note}</td>
-                <td>
-                  <button onClick={() => setTransactionCurrentlyBeingUpdated(transaction)} className="btn btn-primary btn-lg mx-3 my-3">Update</button>
-                  <button onClick={() => { if (window.confirm(`Are you sure you want to delete?`)) deleteTransaction(transaction.transactionId) }} className="btn btn-warning btn-lg">Delete</button>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    )
-  }
-
   function renderTransactionsTable() {
-    if(reloadData){
+    if (reloadData) {
       setReloadData(false);
       getTransactions();
     }
     return (
       <div className="table-responsive mt-4">
-        <TransactionTable data={transactions}/>
+        <TransactionTable data={transactions} 
+        setTransactionCurrentlyBeingUpdated={setTransactionCurrentlyBeingUpdated}
+        deleteTransaction={deleteTransaction} />
       </div>
     )
   }
@@ -117,7 +91,6 @@ export default function App() {
     alert(`Transaction successfully created. After clicking OK, your new transaction of value "${createdTransaction.value}" will show up in the table below.`);
 
     setReloadData(true);
-    renderTransactionsTable();
   }
 
   function onTransactionUpdated(updatedTransaction) {
